Fixed the Quiz Section module, added new features.

+++ QuizMakerDb/Data/Models/Quiz.cs
@@ -21,7 +21,7 @@ namespace QuizMakerDb.Data.Models
 		[StringLength(300)]
 		public string ConclusionDescription { get; set; } = string.Empty!;
 
-        public byte Minutes { get; set; }
+        public Int16 Minutes { get; set; }
 
         public byte Takes { get; set; }
